Welcome to the danger zone
I keep finding myself wanting to use ImmutableArray<T>, but needing a regular ol' T[] all over the place is such a drag (and it doesn't help that there are places that just return T[] which I know nobody's ever going to modify, yet I still need to copy it all to use ImmutableArray<T>).  So to encourage me to use it more, I'm making it possible to convert between T[] and ImmutableArray<T>, without any need to copy the underlying data.

Does this defeat the point of ImmutableArray<T>?  Well... yes and no.  Immutable collections can't be truly immutable in full trust, since we can always just use reflection, and System.Runtime.CompilerServices.Unsafe opens up all kinds of portals to and from Hell (case in point: this...).

The actual point of immutable collections (and especially ImmutableArray<T>) is that they offer promises that are difficult enough to break that if they cannot be considered guarantees, then it's not the consumer's fault.  Kinda like how System.String is supposed to be considered immutable, even though the language even has special support for pinning it and getting a pointer that you can party on all you want...